+++ components/layout/footer/footer.tsx
@@ -2,13 +2,13 @@ import React from "react";
 import Link from "next/link";
 import { FaFacebookF, FaGithub, FaTwitter } from "react-icons/fa";
 import { AiFillInstagram } from "react-icons/ai";
-import { Container } from "./container";
+import { Container } from "../../util/container";
 import { RawRenderer } from "./rawRenderer";
-import { ThemeContext } from "./theme";
-import { Icon } from "./icon";
+import { useTheme } from "..";
+import { Icon } from "../../util/icon";
 
 export const Footer = ({ data, icon, rawData }) => {
-  const theme = React.useContext(ThemeContext);
+  const theme = useTheme();
   const socialIconClasses = "h-7 w-auto";
   const socialIconColorClasses = {
     blue: "text-blue-500 dark:text-blue-400 hover:text-blue-300",
