+++ components/layout/index.tsx
@@ -1,0 +1,2 @@
+export { Layout } from "./layout";
+export { useTheme } from "./theme";
