+++ components/posts/index.tsx
@@ -1,0 +1,1 @@
+export { Posts } from "./posts";
