+++ components/util/section.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { ThemeContext } from "./theme";
+import { useTheme } from "../layout";
 
 export const Section = ({ children, color = "", className = "" }) => {
-  const theme = React.useContext(ThemeContext);
+  const theme = useTheme();
   const sectionColor = {
     default:
       "text-gray-800 dark:text-gray-50 bg-gradient-to-br from-gray-50 dark:from-gray-800 via-transparent to-transparent",
